Current status
- major functionality implemented
- pull-to-refresh implmented
- BUG - pull to refresh leaves the view in a bad state (but not broken)
- started seeing the following:
"Oct 23 00:09:24 02244A-mkim twitter[895] <Error>: CGContextAddPath: invalid context 0x0. This is a serious error. This application, or a library it uses, is using an invalid context  and is thereby contributing to an overall degradation of system stability and reliability. This notice is a courtesy: please fix this problem. It will become a fatal error in an upcoming update."